+++ src/mmds/src/lib.rs
@@ -11,6 +11,7 @@ use serde_json::{Map, Value};
 use std::sync::{Arc, Mutex};
 
 use crate::data_store::{Error as MmdsError, Mmds, MmdsVersion, OutputFormat};
+use crate::token::PATH_TO_TOKEN;
 
 use lazy_static::lazy_static;
 use micro_http::{Body, MediaType, Method, Request, Response, StatusCode, Version};
@@ -120,23 +121,6 @@ fn respond_to_request_mmdsv1(request: Request) -> Response {
 }
 
 fn respond_to_request_mmdsv2(request: Request) -> Response {
-    // Allow only GET requests.
-    // Will allow PUT requests for MMDSv2 in a subsequent commit.
-    match request.method() {
-        Method::Get => respond_to_get_request_checked(request),
-        _ => {
-            let mut response = build_response(
-                request.http_version(),
-                StatusCode::MethodNotAllowed,
-                Body::new("Not allowed HTTP method."),
-            );
-            response.allow_method(Method::Get);
-            response
-        }
-    }
-}
-
-fn respond_to_get_request_checked(request: Request) -> Response {
     // Fetch custom headers from request.
     let token_headers = match TokenHeaders::try_from(request.headers.custom_entries()) {
         Ok(token_headers) => token_headers,
@@ -149,6 +133,24 @@ fn respond_to_get_request_checked(request: Request) -> Response {
         }
     };
 
+    // Allow only GET and PUT requests.
+    match request.method() {
+        Method::Get => respond_to_get_request_checked(request, token_headers),
+        Method::Put => respond_to_put_request(request, token_headers),
+        _ => {
+            let mut response = build_response(
+                request.http_version(),
+                StatusCode::MethodNotAllowed,
+                Body::new("Not allowed HTTP method."),
+            );
+            response.allow_method(Method::Get);
+            response.allow_method(Method::Put);
+            response
+        }
+    }
+}
+
+fn respond_to_get_request_checked(request: Request, token_headers: TokenHeaders) -> Response {
     // Get MMDSv2 token from custom headers.
     let token = match token_headers.x_metadata_token() {
         Some(token) => token,
@@ -163,7 +165,8 @@ fn respond_to_get_request_checked(request: Request) -> Response {
     };
 
     // Validate Mmds V2 token.
-    match MMDS.lock().expect("Poisoned lock").is_valid_token(token) {
+    let is_valid = MMDS.lock().expect("Poisoned lock").is_valid_token(token);
+    match is_valid {
         Ok(true) => respond_to_get_request_unchecked(request),
         Ok(false) => {
             let err_msg = "Mmds V2 token not valid.";
@@ -223,9 +226,54 @@ fn respond_to_get_request_unchecked(request: Request) -> Response {
     }
 }
 
+fn respond_to_put_request(request: Request, token_headers: TokenHeaders) -> Response {
+    let uri = request.uri().get_abs_path();
+    // Sanitize the URI into a strict json path.
+    let json_pointer = sanitize_uri(uri.to_string());
+
+    // Only accept PUT requests towards TOKEN_PATH.
+    if json_pointer != PATH_TO_TOKEN {
+        let error_msg = format!("Resource not found: {}.", uri);
+        return build_response(
+            request.http_version(),
+            StatusCode::NotFound,
+            Body::new(error_msg),
+        );
+    }
+
+    // Get token lifetime value.
+    let ttl_seconds = match token_headers.x_metadata_token_ttl_seconds() {
+        Some(ttl_seconds) => ttl_seconds,
+        None => {
+            let error_msg = "Token time to live value not found.".to_string();
+            return build_response(
+                request.http_version(),
+                StatusCode::BadRequest,
+                Body::new(error_msg),
+            );
+        }
+    };
+
+    // Generate token.
+    match MMDS
+        .lock()
+        .expect("Poisoned lock")
+        .generate_token(ttl_seconds)
+    {
+        Ok(token) => build_response(request.http_version(), StatusCode::OK, Body::new(token)),
+        Err(err) => build_response(
+            request.http_version(),
+            StatusCode::BadRequest,
+            Body::new(err.to_string()),
+        ),
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::token::{MAX_TOKEN_TTL_SECONDS, MIN_TOKEN_TTL_SECONDS};
+    use std::time::Duration;
 
     fn populate_mmds() -> Arc<Mutex<Mmds>> {
         let data = r#"{
@@ -383,12 +431,94 @@ mod tests {
         let actual_response = convert_to_response(request);
         assert_eq!(actual_response, expected_response);
 
-        // Test GET request with invalid token.
-        let request_bytes = b"GET http://169.254.169.254/ HTTP/1.0\r\n\
-                                    X-metadata-token: foo\r\n\r\n";
+        // Test PUT requests.
+        // Test invalid path.
+        let request_bytes = b"PUT http://169.254.169.254/token HTTP/1.0\r\n\
+                                    X-metadata-token-ttl-seconds: 60\r\n\r\n";
         let request = Request::try_from(request_bytes, None).unwrap();
-        let mut expected_response = Response::new(Version::Http10, StatusCode::Unauthorized);
-        expected_response.set_body(Body::new("Mmds V2 token not valid.".to_string()));
+        let mut expected_response = Response::new(Version::Http10, StatusCode::NotFound);
+        expected_response.set_body(Body::new("Resource not found: /token.".to_string()));
+        let actual_response = convert_to_response(request);
+        assert_eq!(actual_response, expected_response);
+
+        // Test invalid lifetime values for token.
+        let invalid_values = [MIN_TOKEN_TTL_SECONDS - 1, MAX_TOKEN_TTL_SECONDS + 1];
+        for invalid_value in invalid_values.iter() {
+            let request_bytes = format!(
+                "PUT http://169.254.169.254/latest/api/token HTTP/1.0\r\n\
+                X-metadata-token-ttl-seconds: {}\r\n\r\n",
+                invalid_value
+            );
+            let request = Request::try_from(request_bytes.as_bytes(), None).unwrap();
+            let mut expected_response = Response::new(Version::Http10, StatusCode::BadRequest);
+            let error_msg = format!(
+                "Invalid time to live value provided for token: {}. \
+                Please provide a value between {} and {}.",
+                invalid_value, MIN_TOKEN_TTL_SECONDS, MAX_TOKEN_TTL_SECONDS
+            );
+            expected_response.set_body(Body::new(error_msg));
+            let actual_response = convert_to_response(request);
+            assert_eq!(actual_response, expected_response);
+        }
+
+        // Test no lifetime value provided for token.
+        let request_bytes = b"PUT http://169.254.169.254/latest/api/token HTTP/1.0\r\n\r\n";
+        let request = Request::try_from(request_bytes, None).unwrap();
+        let mut expected_response = Response::new(Version::Http10, StatusCode::BadRequest);
+        expected_response.set_body(Body::new("Token time to live value not found."));
+        let actual_response = convert_to_response(request);
+        assert_eq!(actual_response, expected_response);
+
+        // Test valid PUT.
+        let request_bytes = b"PUT http://169.254.169.254/latest/api/token HTTP/1.0\r\n\
+                                    X-metadata-token-ttl-seconds: 1\r\n\r\n";
+        let request = Request::try_from(request_bytes, None).unwrap();
+        let actual_response = convert_to_response(request);
+        assert_eq!(actual_response.status(), StatusCode::OK);
+
+        // Test valid GET.
+        let valid_token = String::from_utf8(actual_response.body().unwrap().body).unwrap();
+        let request_bytes = format!(
+            "GET http://169.254.169.254/ HTTP/1.0\r\n\
+            Accept: application/json\r\n\
+            X-metadata-token: {}\r\n\r\n",
+            valid_token
+        );
+        let request = Request::try_from(request_bytes.as_bytes(), None).unwrap();
+        let mut expected_response = Response::new(Version::Http10, StatusCode::OK);
+        let mut body = get_json_data().to_string();
+        body.retain(|c| !c.is_whitespace());
+        expected_response.set_body(Body::new(body));
+        let actual_response = convert_to_response(request);
+        assert_eq!(actual_response, expected_response);
+
+        // Test GET request with invalid tokens.
+        // `valid_token` will become invalid after one second, when it expires.
+        let invalid_token = std::iter::repeat("a").take(58).collect::<String>();
+        let tokens = [invalid_token, valid_token];
+        for token in tokens.iter() {
+            let request_bytes = format!(
+                "GET http://169.254.169.254/ HTTP/1.0\r\n\
+                X-metadata-token: {}\r\n\r\n",
+                token
+            );
+            let request = Request::try_from(request_bytes.as_bytes(), None).unwrap();
+            let mut expected_response = Response::new(Version::Http10, StatusCode::Unauthorized);
+            expected_response.set_body(Body::new("Mmds V2 token not valid.".to_string()));
+            let actual_response = convert_to_response(request);
+            assert_eq!(actual_response, expected_response);
+
+            // Wait for the second token to expire.
+            std::thread::sleep(Duration::from_secs(1));
+        }
+
+        // Test not allowed PATCH HTTP Method.
+        let request_bytes = b"PATCH http://169.254.169.255/ HTTP/1.0\r\n\r\n";
+        let request = Request::try_from(request_bytes, None).unwrap();
+        let mut expected_response = Response::new(Version::Http10, StatusCode::MethodNotAllowed);
+        expected_response.set_body(Body::new("Not allowed HTTP method.".to_string()));
+        expected_response.allow_method(Method::Get);
+        expected_response.allow_method(Method::Put);
         let actual_response = convert_to_response(request);
         assert_eq!(actual_response, expected_response);
     }